+++ client/src/app.styles.css
@@ -3,3 +3,6 @@ body {
   font-family: "Roboto";
   color: white;
 }
+a {
+  text-decoration: none;
+}
