+++ client/src/assets/GridElements.js
@@ -1,0 +1,14 @@
+import styled from "styled-components";
+
+const Div = styled.div`
+  display: flex;
+  margin: 10px 10px 0px 0px;
+  justify-content: flex-end;
+  align-items: center;
+`;
+
+const Span = styled.span`
+  margin: 0px 5px;
+`;
+
+export { Div, Span };
